[#456][style] fix clippy warnings across codebase
- Remove stale #[allow(dead_code)] on spawn_tun_rx/spawn_tun_tx (both
  called from orch.rs)
- Use is_some_and/is_none_or instead of map_or(bool, closure)
- Use io::Error::other() instead of io::Error::new(ErrorKind::Other, ..)
- Use struct update syntax instead of field assignment after default()
- Remove unit-value let binding in helpers test
- Replace useless vec! with array literal in orch test
- Remove unnecessary u16 cast in e2e test

+++ tests/e2e/bareudp.rs
@@ -270,7 +270,7 @@ async fn test_mtu_boundary_drop() {
 
     // Compute ping payload sizes from default MTU.
     // Max payload = MTU - 20 (IP hdr) - 8 (ICMP hdr).
-    let mtu = h3llo::config::default_mtu() as u16;
+    let mtu = h3llo::config::default_mtu();
     let ok_payload = (mtu - 20 - 8).to_string();
     let exceed_payload = mtu.to_string();
